+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+All notable changes to `limesctl` will be documented in this file.
 
 The sections should follow the order `Added`, `Changed`, `Fixed`, `Removed`, and `Deprecated`.
 
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/sapcc/limesctl/compare/v2.0.1...HEAD)
+
+### Changed
+
+- Updated all dependencies to their latest version.
 
 ## 2.0.1 - 2021-10-06
 
@@ -58,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ```
 
   New style option 2, use `-q` shorthand flag for each new quota value:
+
   ```
   limesctl project set -q compute/cores=250 -q compute/ram=20GiB
   ```