+++ CHANGELOG.md
@@ -7,7 +7,13 @@ The sections should follow the order `Added`, `Changed`, `Fixed`, `Removed`, and
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.0.0...HEAD)
+## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.0.1...HEAD)
+
+## 3.0.1 - 2022-08-30
+
+### Changed
+
+- Updated all dependencies to their latest version.
 
 ## 3.0.0 - 2022-03-03